refactor: Movendo testes para pasta exclusiva
Para separar os testes de unidade e os testes de integração, resolvi criar uma pasta de testes onde todos os arquivos de testes ficarão. Isso se aplica para cada app. Agora, para rodar os testes, a partir da pasta com o manage.py deverá rodar o comando: python manage.py test <app-name>/tests.
